+++ internal/service/advancedcluster/model_to_ClusterDescription20240805.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"sort"
 
-	"go.mongodb.org/atlas-sdk/v20250312017/admin"
+	"go.mongodb.org/atlas-sdk/v20250312018/admin"
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/types"