+++ internal/service/flexsnapshot/model.go
@@ -1,7 +1,7 @@
 package flexsnapshot
 
 import (
-	"go.mongodb.org/atlas-sdk/v20250312017/admin"
+	"go.mongodb.org/atlas-sdk/v20250312018/admin"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"