+++ scripts/generate-doc.sh
@@ -34,7 +34,7 @@
 
 set -euo pipefail
 
-TF_VERSION="${TF_VERSION:-"1.14.7"}" # TF version to use when running tfplugindocs. Default: 1.14.7
+TF_VERSION="${TF_VERSION:-"1.14.8"}" # TF version to use when running tfplugindocs. Default: 1.14.8
 TEMPLATE_FOLDER_PATH="${TEMPLATE_FOLDER_PATH:-"templates"}" # PATH to the templates folder. Default: templates