+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ENHANCEMENTS:
 
+* data-source/mongodbatlas_stream_connection: Adds `AzureBlobStorage` connection ([#4357](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/4357))
 * data-source/mongodbatlas_stream_connection: Adds support for `GCPPubSub` connection type with `gcp` attribute ([#4354](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/4354))
+* data-source/mongodbatlas_stream_connections: Adds `AzureBlobStorage` connection ([#4357](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/4357))
 * data-source/mongodbatlas_stream_connections: Adds support for `GCPPubSub` connection type with `gcp` attribute ([#4354](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/4354))
+* resource/mongodbatlas_stream_connection: Adds `AzureBlobStorage` connection ([#4357](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/4357))
 * resource/mongodbatlas_stream_connection: Adds support for `GCPPubSub` connection type with `gcp` attribute ([#4354](https://github.com/mongodb/terraform-provider-mongodbatlas/pull/4354))
 
 BUG FIXES: